% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_update_fn.R
\name{create_test_update_fns}
\alias{create_test_update_fns}
\title{Create Test Update Functions}
\usage{
create_test_update_fns(fn_names, id_arg = "inputId", .package = "shiny")
}
\arguments{
\item{fn_names, fn_name}{A character vector (string) of function names to create wrappers for}

\item{id_arg}{A character string of the argument in `fn_names` that relates to the input argument.
Default is `"inputId"`}

\item{.package}{Character string of the package that `fn_names` exist in.
Default is `"shiny"`}
}
\value{
A named list of function expressions, one for each function supplied in `fn_names`.
}
\description{
Given a set of functions from an R package, create a set of mocked functions that
can be used as bindings to test UI updates within `testServer`.
}
\examples{
create_test_update_fns(
  c("updateSelectInput", "updateTextInput"),
  .package = "shiny"
)

}
