% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_update_fn.R
\name{create_test_update_fns}
\alias{create_test_update_fns}
\title{Create Test Update Functions}
\usage{
create_test_update_fns(fn_names, .package = "shiny")
}
\description{
Create Test Update Functions
}
