% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{splice-operator}
\alias{splice-operator}
\title{Splice Operator !!!}
\value{
The result of calling `do.call(fn_name, ...)`.
}
\description{
See \code{rlang::\link[rlang:splice-operator]{!!!}} for details.
}
\examples{
dfs <- list(mtcars, mtcars)
rlang::inject(rbind(!!!dfs))


}
\keyword{internal}
