% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_input.R
\name{update_shiny_input}
\alias{update_shiny_input}
\title{Update Shiny Input}
\usage{
update_shiny_input(
  id_arg = "inputId",
  value_arg = "value",
  label_arg = "label"
)
}
\description{
When using \code{\link[shiny{testServer}]}
}
