% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_update_fn.R
\name{create_test_update_fns}
\alias{create_test_update_fns}
\title{Create Test Update Functions}
\usage{
create_test_update_fns(fn_names, .package = "shiny")
}
\arguments{
\item{fn_names, fn_name}{A character vector (string) of function names to create wrappers for}

\item{.package}{Character string of the package that `fn_names` exist in}
}
\value{
TODO
}
\description{
TODO
}
\examples{
create_test_update_fns(
  c("updateSelectInput", "updateTextInput"),
  .package = "shiny"
)

}
