% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_shiny_testers.R
\name{use_shiny_testers}
\alias{use_shiny_testers}
\title{Use Shiny Testers}
\usage{
use_shiny_testers(.env = rlang::caller_env())
}
\arguments{
\item{.env}{Environment that defines effect scope. For expert use only.}
}
\value{
TODO
}
\description{
TODO
}
